Rename wellboreItemShapeGenerator
